# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "DownloadJobStartRes": {
        "properties": {
          "job_id": {
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "type": "object"
      },
...
      "FfmpegDownloadStartReq": {
        "type": "object"
      },
...
      "PageResult_StashSceneRow": {
        "properties": {
          "data": {
            "items": {
              "properties": {
                "captions": {
                  "items": {
                    "$ref": "#/components/schemas/StashSceneCaption"
                  },
                  "type": "array"
                },
                "date": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "details": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "files": {
                  "items": {
                    "$ref": "#/components/schemas/StashSceneFile"
                  },
                  "type": "array"
                },
                "id": {
                  "type": "string"
                },
                "last_played_at": {
                  "description": "最后播放时间",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "organized": {
                  "type": [
                    "boolean",
                    "null"
                  ]
                },
                "paths": {
                  "$ref": "#/components/schemas/StashScenePaths"
                },
                "title": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "title",
                "files",
                "paths"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "total": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "type": "object"
      },
...
      "StashDateCriterion": {
        "description": "Stash `DateCriterionInput` / `TimestampCriterionInput`",
        "properties": {
          "modifier": {
            "$ref": "#/components/schemas/StashCriterionModifier"
          },
          "value": {
            "type": "string"
          },
          "value2": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "value",
          "modifier"
        ],
        "type": "object"
      },
...
      "StashEntityKind": {
        "description": "Stash 可搜索实体类型。",
        "enum": [
          "studio",
          "performer",
          "tag"
        ],
        "type": "string"
      },
      "StashEntitySearchItem": {
        "description": "Stash 实体搜索项。",
        "properties": {
          "disambiguation": {
            "description": "演员重名区分信息，其他实体为空。",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "description": "Stash 实体 ID。",
            "type": "string"
          },
          "name": {
            "description": "显示名称。",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "type": "object"
      },
...
      "StashEntitySearchRes": {
        "description": "Stash 实体搜索结果。",
        "properties": {
          "items": {
            "description": "命中的实体列表。",
            "items": {
              "$ref": "#/components/schemas/StashEntitySearchItem"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "StashFilter": {
        "description": "Stash 场景列表查询（分页字段与 ProTable 对齐，服务端映射为 Stash `FindFilterType`）",
        "properties": {
          "direction": {
            "description": "`ASC` 或 `DESC`",
            "type": [
              "string",
              "null"
            ]
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "page_size": {
            "format": "int32",
            "type": "integer"
          },
          "q": {
            "type": [
              "string",
              "null"
            ]
          },
          "sort": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "StashIdentifyFieldOption": {
        "description": "Stash 元数据补全的单个字段策略。",
        "properties": {
          "createMissing": {
            "description": "演员、标签、工作室等关联对象不存在时是否创建。",
            "type": [
              "boolean",
              "null"
            ]
          },
          "field": {
            "description": "字段名，与 Stash `IdentifyFieldOptionsInput.field` 保持一致。",
            "type": "string"
          },
          "strategy": {
            "$ref": "#/components/schemas/StashIdentifyFieldStrategy",
            "description": "字段写入策略。"
          }
        },
        "required": [
          "field",
          "strategy"
        ],
        "type": "object"
      },
      "StashIdentifyFieldStrategy": {
        "description": "Stash 元数据识别字段写入策略。",
        "enum": [
          "IGNORE",
          "MERGE",
          "OVERWRITE"
        ],
        "type": "string"
      },
      "StashIdentifySource": {
        "description": "Stash 元数据识别来源，可指向 StashBox 端点或本地 scraper。",
        "properties": {
          "scraper_id": {
            "description": "本地 scraper ID，例如 ThePornDB scraper 的 ID。",
            "type": [
              "string",
              "null"
            ]
          },
          "stash_box_endpoint": {
            "description": "StashBox GraphQL 端点，例如 `https://stashdb.org/graphql`。",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "StashIntCriterion": {
        "description": "Stash `IntCriterionInput`",
        "properties": {
          "modifier": {
            "$ref": "#/components/schemas/StashCriterionModifier"
          },
          "value": {
            "format": "int32",
            "type": "integer"
          },
          "value2": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "value",
          "modifier"
        ],
        "type": "object"
      },
...
      "StashSceneCaption": {
        "properties": {
          "caption_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "language_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "local_path": {
            "description": "根据 `stash_config.path_mappings` 映射出的本服务本地字幕路径。",
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "description": "Stash 视角推导出的字幕文件路径。",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "StashSceneFile": {
        "properties": {
          "basename": {
            "type": "string"
          },
          "duration": {
            "description": "影片时长，单位为秒。",
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "local_path": {
            "description": "根据 `stash_config.path_mappings` 映射出的本服务本地路径。",
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "basename"
        ],
        "type": "object"
      },
      "StashSceneFilterType": {
        "description": "Stash GraphQL `SceneFilterType`（字段与官方 schema 对齐，未建模的关联过滤器以 JSON 透传）",
        "properties": {
          "AND": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashSceneFilterType"
              }
            ]
          },
          "NOT": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashSceneFilterType"
              }
            ]
          },
          "OR": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashSceneFilterType"
              }
            ]
          },
          "audio_codec": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "bitrate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "captions": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "checksum": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "created_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashDateCriterion"
              }
            ]
          },
          "custom_fields": {
            "items": {
              "$ref": "#/components/schemas/StashCustomFieldCriterion"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "date": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashDateCriterion"
              }
            ]
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "director": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "duplicated": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashDuplicationCriterion"
              }
            ]
          },
          "duration": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "file_count": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "files_filter": {},
          "framerate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "galleries": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashMultiCriterion"
              }
            ]
          },
          "galleries_filter": {
            "description": "关联实体过滤器（结构体庞大，按需透传 Stash 原生 JSON）"
          },
          "groups": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashHierarchicalMultiCriterion"
              }
            ]
          },
          "groups_filter": {},
          "has_markers": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "interactive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "interactive_speed": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "is_missing": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_played_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashDateCriterion"
              }
            ]
          },
          "markers_filter": {},
          "movies": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashMultiCriterion"
              }
            ]
          },
          "movies_filter": {},
          "o_counter": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "organized": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "orientation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashOrientationCriterion"
              }
            ]
          },
          "oshash": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "path": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "performer_age": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "performer_count": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "performer_favorite": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "performer_tags": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashHierarchicalMultiCriterion"
              }
            ]
          },
          "performers": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashMultiCriterion"
              }
            ]
          },
          "performers_filter": {},
          "phash": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "phash_distance": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashPhashDistanceCriterion"
              }
            ]
          },
          "play_count": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "play_duration": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "rating100": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "resolution": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashResolutionCriterion"
              }
            ]
          },
          "resume_time": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "stash_id_count": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "stash_id_endpoint": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIdCriterion"
              }
            ]
          },
          "stash_ids_endpoint": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIdsCriterion"
              }
            ]
          },
          "studios": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashHierarchicalMultiCriterion"
              }
            ]
          },
          "studios_filter": {},
          "tag_count": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashIntCriterion"
              }
            ]
          },
          "tags": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashHierarchicalMultiCriterion"
              }
            ]
          },
          "tags_filter": {},
          "title": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "updated_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashDateCriterion"
              }
            ]
          },
          "url": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          },
          "video_codec": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashStringCriterion"
              }
            ]
          }
        },
        "type": "object"
      },
      "StashSceneListReq": {
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/StashFilter"
          },
          "scene_filter": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StashSceneFilterType",
                "description": "对应 GraphQL `SceneFilterType`"
              }
            ]
          },
          "scene_ids": {
            "description": "对应 GraphQL `scene_ids`",
            "items": {
              "format": "int32",
              "type": "integer"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "filter"
        ],
        "type": "object"
      },
      "StashSceneMetadataCompleteReq": {
        "description": "Stash 场景元数据补全请求。",
        "properties": {
          "field_options": {
            "description": "字段策略；不传时只补空字段并合并关联实体。",
            "items": {
              "$ref": "#/components/schemas/StashIdentifyFieldOption"
            },
            "type": "array"
          },
          "scene_ids": {
            "description": "要补全的 Stash 场景 ID；不传时自动查询所有已整理但标题为空的场景。",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "set_cover_image": {
            "description": "是否设置封面图。",
            "type": "boolean"
          },
          "set_organized": {
            "description": "是否将场景标记为已整理。",
            "type": "boolean"
          },
          "skip_multiple_matches": {
            "description": "多个匹配结果时跳过，降低误写入风险。",
            "type": "boolean"
          },
          "sources": {
            "description": "识别来源，按顺序尝试；不传时默认 StashDB -> ThePornDB。",
            "items": {
              "$ref": "#/components/schemas/StashIdentifySource"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "StashSceneMetadataCompleteRes": {
        "description": "Stash 场景元数据补全响应。",
        "properties": {
          "job_id": {
            "description": "Stash 创建的元数据识别任务 ID；没有命中场景时为空。",
            "type": [
              "string",
              "null"
            ]
          },
          "scene_count": {
            "description": "本次提交的场景数量。",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "scene_count"
        ],
        "type": "object"
      },
      "StashScenePaths": {
        "properties": {
          "preview": {
            "type": "string"
          },
          "screenshot": {
            "type": "string"
          }
        },
        "required": [
          "screenshot",
          "preview"
        ],
        "type": "object"
      },
...
      "StashStringCriterion": {
        "description": "Stash `StringCriterionInput`",
        "properties": {
          "modifier": {
            "$ref": "#/components/schemas/StashCriterionModifier"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "value",
          "modifier"
        ],
        "type": "object"
      },
...
      "WhisperDownloadStartReq": {
        "properties": {
          "model_id": {
            "type": "string"
          }
        },
        "required": [
          "model_id"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "media-admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/setup/ffmpeg/download": {
      "post": {
        "operationId": "downloadFfmpegSetup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FfmpegDownloadStartReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobStartRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "setup"
        ]
      }
    },
...
    "/api/setup/whisper/download": {
      "post": {
        "operationId": "downloadWhisperSetup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhisperDownloadStartReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobStartRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "setup"
        ]
      }
    },
...
    "/api/stash/entities/search": {
      "get": {
        "operationId": "searchEntitiesStash",
        "parameters": [
          {
            "description": "实体类型。",
            "in": "query",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/StashEntityKind"
            }
          },
          {
            "description": "搜索关键词。",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "返回数量。",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StashEntitySearchRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "stash"
        ]
      }
    },
    "/api/stash/scenes/list": {
      "post": {
        "operationId": "listScenesStash",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StashSceneListReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult_StashSceneRow"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "stash"
        ]
      }
    },
    "/api/stash/scenes/metadata/complete": {
      "post": {
        "operationId": "completeSceneMetadataStash",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StashSceneMetadataCompleteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StashSceneMetadataCompleteRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "stash"
        ]
      }
    },
...
  },
...
}